refactor proposal delivery events router
